Module 1 !practice

+++ src/module1/_7math/E01Random.java
@@ -4,7 +4,7 @@ public class E01Random {
 	//Exercise 1: Run this program several times and observe
 	//what gets printed
 	public static void main(String[] args) {
-		System.out.println(Math.random());
+		System.out.println((int)Math.random());
 		System.out.println(Math.random());
 	}
 }
